inceptionv3_transfer/requirements.txt
gast==0.3.3
numpy==1.18.5
Pillow==7.1.2
protobuf==3.12.2
requests==2.23.0
tensorflow==2.2.0

